Move TimeSpanExtensions.cs to TwitchDownloaderCore.Extensions
